debug: let it always return 0 to avoid unexpect behavior
The bug is from this line:
[[ "$1" == */ ]] && [ -d "$1" ] && dal $LINENO mrFILE "$1" && return 2
I found when --debug the return value is 2 as expected,
when no --debug it will return unexpected 0.

My analysis:
"&& return 2" need "dal $LINENO mrFILE $1" to return 0,
and dal's return value is debug's return value:
dal() { eval "$2=\"$3\""; debug "$1: $2=\"$3\""; }
If debug does not explicitly return 0, like:
debug() { [ "$debug" = true ] && >&2 echo "$@"; }
it's return value is the last command's return value.
When $debug==true, the last command is echo, so the return value is 0;
When $debug!=true, the last command is [ "$debug" = true ],
and its return value is 1 for different value.

So, the original debug function has side effect,
to eliminate it, just let is explicitly return 0.